all functionalities done

+++ frontend/src/pages/MyTasks/TaskDetails.jsx
@@ -1,35 +1,53 @@
 import { useMyTasksContext } from "../../hooks/useTaskContext";
+import { useAuthContext } from "../../hooks/useAuthContext";
 
-import formatDistanceToNow from "date-fns/formatDistanceToNow"
+import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
-const TaskDetails = ({task}) => {
-    const {dispatch} = useMyTasksContext()
+const TaskDetails = ({ task }) => {
+  const { dispatch } = useMyTasksContext();
+  const { user } = useAuthContext();
 
-    const handleClick = async () => {
-        try {
-            const response = await fetch(`http://localhost:5000/api/myTasks/${task._id}`, {
-                method: 'DELETE',
-            });
+  const handleClick = async () => {
+    if (!user) {
+      return;
+    }
 
-            if (!response.ok) {
-                throw new Error('Failed to delete task');
-            }
-
-            const deletedTask = await response.json();
-            dispatch({ type: 'DELETE_TASK', payload: deletedTask });
-        } catch (error) {
-            console.error('Error deleting task:', error);
+    try {
+      const response = await fetch(
+        `http://localhost:5000/api/myTasks/${task._id}`,
+        {
+          method: "DELETE",
+          headers: {
+            Authorization: `Bearer ${user.token}`,
+          },
         }
-    };
-      
+      );
 
-    return(
-        <div style={{border:"1px solid black", padding:"1rem"}}>
-            <p>{task.title}</p>
-            <p>{formatDistanceToNow(new Date(task.createdAt), {addSuffix: true})}</p>
-            <span onClick={handleClick} style={{cursor:"pointer", backgroundColor:"red"}}>Delete</span>
-        </div>
-    )
-}
+      if (!response.ok) {
+        throw new Error("Failed to delete task");
+      }
+
+      const deletedTask = await response.json();
+      dispatch({ type: "DELETE_TASK", payload: deletedTask });
+    } catch (error) {
+      console.error("Error deleting task:", error);
+    }
+  };
+
+  return (
+    <div style={{ border: "1px solid black", padding: "1rem" }}>
+      <p>{task.title}</p>
+      <p>
+        {formatDistanceToNow(new Date(task.createdAt), { addSuffix: true })}
+      </p>
+      <span
+        onClick={handleClick}
+        style={{ cursor: "pointer", backgroundColor: "red" }}
+      >
+        Delete
+      </span>
+    </div>
+  );
+};
 
 export default TaskDetails;
